Updated the return type of GetCharInput(), fixed a bug in the IsPlayable() method, and fixed a couple compiler warnings.

+++ unoApp/UnoApp.cpp
@@ -5,7 +5,7 @@ int main()
 {
     //
     // Seed random number generator.
-    std::srand(std::time(0));
+    std::srand((unsigned int)std::time(0));
 
     //FeatureTest(); // Comment out if desired.
     //KeyInputTest();
